requirements.txt
google-generativeai
streamlit
pandas
reportlab
python-dotenv
pytest
streamlit-mic-recorder
SpeechRecognition
pydub
opencv-python
mediapipe
numpy
